+++ CHANGELOG.md
@@ -26,6 +26,12 @@ All notable changes to this project will be documented in this file.
 
 - split tests utils create change event function into separate functions for input and textarea element types
 
+### Removed
+
+#### Utils
+
+- create change event for both input and textarea combined
+
 ## [0.5.2] - 2025-12-08
 
 - updated dependencies